// ClientApp/app/Models/UnderProcessMandateOutWard/underProcessMandateOutWard.ts
export class UnderProcessMandateOutWard {
    dataList:any;
    sponsorbankcode: string;
    name: string;
    CorpName: string;
   // EntityID: string
    IsMobileData: Boolean;
    createdon: string;
    IsScan: string;
    JPGPath: string;
    TIPPath: string;
    IsPrint: string;
    mandateid: any;
    BankName: string;
    status: string;
    Amount: string;
    IFSC: string;
    Code: string;
 // public string BankName { get; set; }
    DateOnMandate: string;
    AcNo: string;
    Refrence1: string;
    FromDate: string;
    Customer1: string; //  Customer1
    debittype: any;
    Frequency: any;
    ToDebit: any;
    UMRN: any;

    updatedon: any;
    CreatedBy: any;
   
}